day 8 problem 2 solved
learned more about different iterator methods, and implemented
a function that takes an iterator as an argument.

In rust unlike C++11, you don't take an iterator object,
you take a generic argument which has to satisfy the bound that it
be an interator.  You can also constrain the thing that _it_ it
generic in with a where clause.  Cool!